chore(build.js): add --watch

+++ scripts/build.js
@@ -34,7 +34,7 @@ const vitePluginDts = () => {
   return plugin
 }
 
-export async function viteBuild(scopes, copyToPath) {
+export async function viteBuild(scopes, copyToPath, watch) {
   const packages = await getPackages()
   const localExternal = packages.map((pkg) => pkg.name)
   const external = [...publicExternal, ...localExternal, ...nodeExternal]
@@ -50,7 +50,8 @@ export async function viteBuild(scopes, copyToPath) {
         emptyOutDir: true,
         lib: { entry, name, fileName, formats: ['umd', 'es'] },
         rollupOptions: { external },
-        sourcemap: true
+        sourcemap: true,
+        watch: watch ? {} : undefined
       },
       configFile: false,
       root,
@@ -61,28 +62,35 @@ export async function viteBuild(scopes, copyToPath) {
           libraryName: 'antd',
           libraryDirectory: 'es',
           style: 'css'
-        })
+        }),
+        {
+          name: 'vite-plugin-vact-builder',
+          closeBundle: () => {
+            if (copyToPath) {
+              const distDir = path.resolve(`${copyToPath}/node_modules/${name}`)
+              fs.copySync(root, distDir)
+            }
+          }
+        }
       ]
+    }).then((result) => {
+      console.log(chalk.greenBright(`${name} Builded! \n`))
+      return result
     })
-      .then(() =>
-        fs.copySync(root, path.resolve(`${copyToPath}/node_modules/${name}`))
-      )
-      .then((result) => {
-        console.log(chalk.greenBright(`${name} Builded! \n`))
-        return result
-      })
   })
 
-  filterPackages(packages, [
-    '@v-act/schema-types',
-    '@v-act/v3dev-component-watcher',
-    '@v-act/widget-context',
-    '@v-act/widget-utils'
-  ]).forEach((pkg) => {
-    const distDir = path.resolve(`${copyToPath}/node_modules/${pkg.name}/`)
-    fs.rmSync(distDir, { force: true, recursive: true })
-    fs.copySync(pkg.location, distDir)
-    const nodemodules = path.resolve(`${distDir}/node_modules/`)
-    fs.rmSync(nodemodules, { recursive: true, force: true })
-  })
+  if (copyToPath) {
+    filterPackages(packages, [
+      '@v-act/schema-types',
+      '@v-act/v3dev-component-watcher',
+      '@v-act/widget-context',
+      '@v-act/widget-utils'
+    ]).forEach((pkg) => {
+      const distDir = path.resolve(`${copyToPath}/node_modules/${pkg.name}/`)
+      fs.rmSync(distDir, { force: true, recursive: true })
+      fs.copySync(pkg.location, distDir)
+      const nodemodules = path.resolve(`${distDir}/node_modules/`)
+      fs.rmSync(nodemodules, { recursive: true, force: true })
+    })
+  }
 }
